+++ src/Core/Data/Dto/Connection/RejectFollowRequestRequestDto.cs
@@ -3,6 +3,6 @@ namespace GamaEdtech.Data.Dto.Connection
     public sealed class RejectFollowRequestRequestDto
     {
         public required long Id { get; set; }
-        public required int UserId { get; set; }
+        public required long UserId { get; set; }
     }
 }
