+++ src/Core/Data/Dto/Connection/ToggleSubscriptionRequestDto.cs
@@ -2,7 +2,7 @@ namespace GamaEdtech.Data.Dto.Connection
 {
     public sealed class ToggleSubscriptionRequestDto
     {
-        public required int ProfileId { get; set; }
-        public required int UserId { get; set; }
+        public required long ProfileId { get; set; }
+        public required long UserId { get; set; }
     }
 }
