+++ src/Core/Data/Dto/Connection/UnFollowRequestDto.cs
@@ -2,8 +2,8 @@ namespace GamaEdtech.Data.Dto.Connection
 {
     public sealed class UnFollowRequestDto
     {
-        public required int ProfileId { get; set; }
-        public required int UserId { get; set; }
+        public required long ProfileId { get; set; }
+        public required long UserId { get; set; }
         public required bool TwoWayRevoke { get; set; }
     }
 }
