+++ src/Core/Data/Dto/Game/SpendPointsRequestDto.cs
@@ -2,7 +2,7 @@ namespace GamaEdtech.Data.Dto.Game
 {
     public sealed class SpendPointsRequestDto
     {
-        public required int UserId { get; set; }
+        public required long UserId { get; set; }
         public required long Points { get; set; }
     }
 }
