+++ src/Core/Data/Dto/Identity/ManageAvatarRequestDto.cs
@@ -2,7 +2,7 @@ namespace GamaEdtech.Data.Dto.Identity
 {
     public sealed class ManageAvatarRequestDto
     {
-        public required int UserId { get; set; }
+        public required long UserId { get; set; }
 
         public required string? Avatar { get; set; }
     }
