+++ src/Core/Data/Dto/Identity/ProfileSettingsDto.cs
@@ -17,7 +17,7 @@ namespace GamaEdtech.Data.Dto.Identity
         public int? Board { get; set; }
         public int? Grade { get; set; }
         public int? Group { get; set; }
-        public int? CoreId { get; set; }
+        public long? CoreId { get; set; }
         public string? Avatar { get; set; }
         public string? UserName { get; set; }
         public string? WalletId { get; set; }
