+++ src/Core/Data/Dto/Message/MessagesRequestDto.cs
@@ -4,8 +4,8 @@ namespace GamaEdtech.Data.Dto.Message
 
     public sealed class MessagesRequestDto
     {
-        public required int ConnectionId { get; set; }
-        public required int UserId { get; set; }
+        public required long ConnectionId { get; set; }
+        public required long UserId { get; set; }
         public PagingDto? PagingDto { get; set; }
     }
 }
