+++ src/Domain/Entity/SchoolComment.cs
@@ -12,7 +12,7 @@ namespace GamaEdtech.Domain.Entity
     using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
     [Table(nameof(SchoolComment))]
-    public class SchoolComment : VersionableEntity<ApplicationUser, int, int?>, IEntity<SchoolComment, long>, ISchoolId
+    public class SchoolComment : VersionableEntity<ApplicationUser, long, long?>, IEntity<SchoolComment, long>, ISchoolId
     {
         [System.ComponentModel.DataAnnotations.Key]
         [Column(nameof(Id), DataType.Long)]
