+++ src/Domain/Entity/Ticket.cs
@@ -23,8 +23,8 @@ namespace GamaEdtech.Domain.Entity
         [Required]
         public long Id { get; set; }
 
-        [Column(nameof(UserId), DataType.Int)]
-        public int? UserId { get; set; }
+        [Column(nameof(UserId), DataType.Long)]
+        public long? UserId { get; set; }
         public ApplicationUser? User { get; set; }
 
         [Column(nameof(CreationDate), DataType.DateTimeOffset)]
