+++ src/Presentation/ViewModel/Message/MessageConnectionResponseViewModel.cs
@@ -2,7 +2,7 @@ namespace GamaEdtech.Presentation.ViewModel.Message
 {
     public sealed class MessageConnectionResponseViewModel
     {
-        public int ConnectionId { get; set; }
+        public long ConnectionId { get; set; }
         public string? ConnectionName { get; set; }
         public int UnreadCount { get; set; }
     }
