+++ src/Presentation/ViewModel/School/SchoolImageInfoViewModel.cs
@@ -4,7 +4,7 @@ namespace GamaEdtech.Presentation.ViewModel.School
     {
         public long Id { get; set; }
         public Uri? FileUri { get; set; }
-        public int CreationUserId { get; set; }
+        public long CreationUserId { get; set; }
         public string? CreationUser { get; set; }
         public string? TagName { get; set; }
         public string? TagIcon { get; set; }
